[language] Add plugin support for languages
